Remove dynamic imports with variable
Summary:
In version 0.2.4, we utilized rollup/plugin-dynamic-import-vars to handle variables in dynamic import paths. While this approach successfully generated bundled builds for CDN distribution, it required users who installed the package from NPM to also add rollup/plugin-dynamic-import-vars as a development dependency, which was not ideal.

With version 0.2.5, we have removed this dependency. Instead, we now use a script that scans the capture folder to generate a registry file. This file facilitates the dynamic import of all environments at build time, streamlining the process and eliminating the need for additional dependencies.

Reviewed By: zjm-meta

Differential Revision:
D69999311

Privacy Context Container: L1233623

fbshipit-source-id: 5f91da91efaac09ab2391bccb68c5be8d10a85cd