Move DOMPointReadOnly to test directory
Summary: DOMPointReadOnly is widely available across mainstream browser clients. However it's not available in the node environment, hence the jest tests can't run. I've relocated the polyfill implementation of DOMPointReadOnly into the test directory, so that it won't be included in the library build, but will enable the tests to run.

Reviewed By: cabanier

Differential Revision:
D75191941

Privacy Context Container: L1233623

fbshipit-source-id: 505a1c276e077cc1118019a8ad79489d825c6875